DHFPROD-4519: Add SearchManager test
